@@ -14,13 +14,13 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.2.0</version>
+        <version>3.2.5</version>
         <relativePath/>
     </parent>
 
     <properties>
         <!-- toolchain versions -->
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <maven.version>3.8.5</maven.version>
 
         <maven.compiler.source>${java.version}</maven.compiler.source>
