@@ -1,5 +1,6 @@
 package jspeccy;
 
+import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -14,7 +15,11 @@ public class JSpeccyApplication {
 
     public static void main(String[] args) {
 
-        new SpringApplicationBuilder(JSpeccyApplication.class).headless(false).run(args);
+        new SpringApplicationBuilder(JSpeccyApplication.class)
+                // Do NOT start an embedded web server in the application when using SpringBoot web dependencies
+                .web(WebApplicationType.NONE)
+                .headless(false)
+                .run(args);
     }
 
 }
