@@ -6,25 +6,30 @@
 
 package snapshots;
 
+import lombok.NonNull;
+
 import java.io.File;
 
 /**
- *
  * @author jsanchez
  */
 public class SnapshotFactory {
-    public static SnapshotFile getSnapshot(File file) {
-        String name = file.getName().toLowerCase();
-        switch (name.substring(name.lastIndexOf("."), name.length())) {
-            case ".sna":
-                return new SnapshotSNA();
-            case ".z80":
-                return new SnapshotZ80();
-            case ".szx":
-                return new SnapshotSZX();
-            case ".sp":
-                return new SnapshotSP();
-        }
-        throw new IllegalArgumentException("No such snapshot format");
+
+    private SnapshotFactory() {
+        // This private constructor is intended to hide the implicit public constructor
     }
+
+    public static SnapshotFile getSnapshot(@NonNull File file) {
+
+        String name = file.getName().toLowerCase();
+        String format = name.substring(name.lastIndexOf("."));
+        return switch (format) {
+            case ".sna" -> new SnapshotSNA();
+            case ".z80" -> new SnapshotZ80();
+            case ".szx" -> new SnapshotSZX();
+            case ".sp" -> new SnapshotSP();
+            default -> throw new IllegalArgumentException(String.format("Unknown snapshot format: '%s'", format));
+        };
+    }
+
 }
